tomatoes_backend/requirements.txt
flask
tensorflow
keras
pillow
pymongo
python-dotenv
inference-sdk
